requirements.txt
streamlit
pandas
numpy
python-dotenv
tweepy
transformers
torch
jax
tensorflow
fugashi
ipadic